go/internal/pkg/day20/day20_test.go
/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
package aoc2020

import "testing"

func Test_reverse(t *testing.T) {
	type args struct {
		s string
	}
	tests := []struct {
		name string
		args args
		want string
	}{
		{"just reverse", args{s: "ABC D"}, "D CBA"},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := reverse(tt.args.s); got != tt.want {
				t.Errorf("reverse() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_8Edges(t *testing.T) {
	type args struct {
		s []string
	}
	tests := []struct {
		name string
		args args
		want []string
	}{
		{"just calculate", args{s: []string{"AB", "CD"}}, []string{"AB", "BD", "DC", "CA", "BA", "DB", "CD", "AC"}},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := calculate8Edges(tt.args.s); !equal(got, tt.want) {
				t.Errorf("calculate8Edges() = %v, want %v", got, tt.want)
			}
		})
	}
}

func equal(a1 []string, a2 []string) bool {
	if len(a1) != len(a2) {
		return false
	}
	for idx, v := range a1 {
		if v != a2[idx] {
			return false
		}
	}
	return true
}
